Allow redis-server version to be parameterized
Allow redis-server version to be parameterized, unfortunately when I try
to parameterize the `vars/main.yml` file and use
` with_items: redis_ubuntu_pkg` the `redis_server_version` is not passed
to the `if else` statement logic. Therefore I have moved the package
definition to the `tasks/main.yml` file.